entity_layout_actions_menu: declare links only once
to avoid the tedious work of comparing what is in large or small mode.

Additionnally, removing unnecessary on:click event listeners:
the job is already done by the Link component call to app.navigateAndLoad
